whatpm/t/tree-test-1.dat

#data
<!DOCTYPE HTML><html title="t"><head><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   title="t"
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head title="t"><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     title="t"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head><title class="t"></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|       class="t"
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body><title class="a"></title></body></html>
#errors
title in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <title>
|       class="a"

#data
<!DOCTYPE HTML><html><head></head><body></body><title class="a"></title></html>
#errors
data after body
title in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <title>
|       class="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body title="t"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     title="t"

#data
<!DOCTYPE HTML><html><head><title></title><meta name="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <meta>
|       name="a"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><meta name="a"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <meta>
|       name="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><meta name="a"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <meta>
|       name="a"

#data
<!DOCTYPE HTML><html><head><title></title><link rel="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <link>
|       rel="a"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><link rel="a"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <link>
|       rel="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><link rel="a"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <link>
|       rel="a"

#data
<!DOCTYPE HTML><html><head><title></title><base href="http://"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <base>
|       href="http://"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><base href="http://"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <base>
|       href="http://"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><base href="http://"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <base>
|       href="http://"

#data
<!DOCTYPE HTML><p><b><p>X
#errors
<b> is not closed before second <p>
<b> is not closed before end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <p>
|       <b>
|         "X"

#data
<!DOCTYPE HTML><b><p>
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>

#data
<!DOCTYPE HTML><b><p>X
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>
|         "X"

#data
<!DOCTYPE HTML><ins><p>XX</ins>YYY
#errors
1,32,not closed:p
1,34,not closed:ins
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ins>
|       <p>
|         "XXYYY"

## when script is disabled
#data
<!DOCTYPE HTML><noscript>aaa<noscript>bbb</noscript>ccc</noscript>
#errors
in noscript:#character
unmacthed noscript end tag
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     "aaa"
|     <noscript>
|       "bbb"
|     "ccc"

## when script is disabled
#data
<!DOCTYPE HTML><body><noscript>aaa<noscript>bbb</noscript>ccc</noscript>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noscript>
|       "aaa"
|       <noscript>
|         "bbb"
|       "ccc"

#data
<!DOCTYPE HTML><iframe>
#errors
no </iframe>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>

#data
<!DOCTYPE HTML><iframe></iframe>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>

#data
<!DOCTYPE HTML><iframe>a</iframe>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>
|       "a"

#data
<!DOCTYPE HTML><noframes>
#errors
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes>a</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>
|       "a"

#data
<!DOCTYPE HTML><noframes>
#errors
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes>a</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>
|       "a"

#data
<!DOCTYPE HTML><xmp>
#errors
no </xmp>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>

#data
<!DOCTYPE HTML><xmp></xmp>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>

#data
<!DOCTYPE HTML><xmp>a</xmp>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>
|       "a"

#data
<!DOCTYPE HTML><html><head><title></title><style></style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title><style type="text/css"></style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|       type="text/css"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title><style type="text/css">p {}</style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|       type="text/css"
|       "p {}"
|   <body>

#data
<!DOCTYPE HTML><html><style type="text/css"></style><title></title></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <style>
|       type="text/css"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><style type="text/css"></style><title></title><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <style>
|       type="text/css"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><title></title><body><p>a<style type="text/css"></style>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>
|       "a"
|       <style>
|         type="text/css"

#data
<!DOCTYPE HTML><form><form></form>
#errors
form in form
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>

#data
<!DOCTYPE HTML><ul><li>a<li>c</ul>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         "a"
|       <li>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><ul><li>a<div>b<li>c</ul>d
#errors
div closed by li
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         "a"
|         <div>
|           "b"
|       <li>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<dt>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<dd>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<dt>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<dd>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<div>b<dt>c</dl>d
#errors
div closed by dt
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|         <div>
|           "b"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<div>b<dd>c</dl>d
#errors
div closed by dd
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|         <div>
|           "b"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<div>b<dt>c</dl>d
#errors
div closed by dt
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|         <div>
|           "b"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<div>b<dd>c</dl>d
#errors
div closed by dd
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|         <div>
|           "b"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><body><form><div>a</form>b
#errors
div not closed before </form>
div not closed before EOF
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <div>
|         "ab"

#data
<!DOCTYPE HTML><ul></body>
#errors
ul not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>

#data
<!DOCTYPE HTML><ul><li></body>
#errors
ul not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>

#data
<!DOCTYPE HTML><dl></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>

#data
<!DOCTYPE HTML><dl><dt></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>

#data
<!DOCTYPE HTML><dl><dd></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>

#data
<!DOCTYPE HTML><p></body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><table><tr><td></body>
#errors
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><table><tr><th></body>
#errors
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<!DOCTYPE HTML><table><thead><tr><td></body>
#errors
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><isindex>
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex action="http://test/">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       action="http://test/"
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex prompt="Input:">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "Input:"
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex name="textarea">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex type="checkbox">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|             type="checkbox"
|       <hr>

#data
<!DOCTYPE HTML><form><isindex></form>
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>

#data
<!DOCTYPE HTML><h1><a><h2>
#errors
h1 not closed
a not closed
h2 not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <a>
|         <h2>

#data
  <p>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>

#data
  a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     "a"

#data
<!---->  <p>
#errors
no DOCTYPE
#document
| <!--  -->
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML>  <p>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
